Split up structure
