+++ src/service/lootDegradation.ts
@@ -2,6 +2,7 @@ import type { BotContext } from "@/context.js";
 
 import * as time from "@/utils/time.js";
 import * as lootService from "@/service/loot.js";
+import { LootKindId } from "@/service/lootData.js";
 import log from "@log";
 
 export async function degradeItems(_context: BotContext) {
@@ -9,7 +10,7 @@ export async function degradeItems(_context: BotContext) {
 
     const now = Date.now();
     const maxKebabAge = time.days(3);
-    const kebabs = await lootService.getLootsByKindId(lootService.LootKindId.DOENER);
+    const kebabs = await lootService.getLootsByKindId(LootKindId.DOENER);
 
     for (const k of kebabs) {
         const itemAge = now - new Date(k.claimedAt).getTime();
@@ -17,10 +18,7 @@ export async function degradeItems(_context: BotContext) {
             continue;
         }
 
-        const fridges = await lootService.getUserLootsByTypeId(
-            k.winnerId,
-            lootService.LootKindId.KUEHLSCHRANK,
-        );
+        const fridges = await lootService.getUserLootsByTypeId(k.winnerId, LootKindId.KUEHLSCHRANK);
 
         if (fridges.length > 0) {
             // user has a fridge, don't verschimmel döner
@@ -32,7 +30,7 @@ export async function degradeItems(_context: BotContext) {
             {
                 displayName: "Verschimmelter Döner",
                 description: "Du hättest ihn früher essen sollen",
-                lootKindId: lootService.LootKindId.VERSCHIMMELTER_DOENER,
+                lootKindId: LootKindId.VERSCHIMMELTER_DOENER,
                 usedImage: null,
                 winnerId: k.winnerId,
                 claimedAt: k.claimedAt,
